refactor some stuff into service/core and service/start_node
