+++ cumulus/polkadot-parachain/src/service/core/mod.rs
@@ -1,0 +1,2 @@
+pub mod lookahead_aura_consensus;
+pub mod relay_chain_consensus;
